Part 148

+++ src/api/services/ListingItemTemplateService.ts
@@ -10,6 +10,7 @@ import { ItemInformationService } from './ItemInformationService';
 import { PaymentInformationService } from './PaymentInformationService';
 import { MessagingInformationService } from './MessagingInformationService';
 import { CryptocurrencyAddressService } from './CryptocurrencyAddressService';
+import { ListingItemObjectService } from './ListingItemObjectService';
 import { ListingItemTemplate } from '../models/ListingItemTemplate';
 import { ListingItemTemplateCreateRequest } from '../requests/ListingItemTemplateCreateRequest';
 import { ListingItemTemplateUpdateRequest } from '../requests/ListingItemTemplateUpdateRequest';
@@ -20,6 +21,8 @@ import { PaymentInformationCreateRequest } from '../requests/PaymentInformationC
 import { PaymentInformationUpdateRequest } from '../requests/PaymentInformationUpdateRequest';
 import { MessagingInformationCreateRequest } from '../requests/MessagingInformationCreateRequest';
 import { MessagingInformationUpdateRequest } from '../requests/MessagingInformationUpdateRequest';
+import { ListingItemObjectCreateRequest } from '../requests/ListingItemObjectCreateRequest';
+import { ListingItemObjectUpdateRequest } from '../requests/ListingItemObjectUpdateRequest';
 
 export class ListingItemTemplateService {
 
@@ -31,6 +34,7 @@ export class ListingItemTemplateService {
         @inject(Types.Service) @named(Targets.Service.PaymentInformationService) public paymentInformationService: PaymentInformationService,
         @inject(Types.Service) @named(Targets.Service.MessagingInformationService) public messagingInformationService: MessagingInformationService,
         @inject(Types.Service) @named(Targets.Service.CryptocurrencyAddressService) public cryptocurrencyAddressService: CryptocurrencyAddressService,
+        @inject(Types.Service) @named(Targets.Service.ListingItemObjectService) public listingItemObjectService: ListingItemObjectService,
         @inject(Types.Core) @named(Core.Logger) public Logger: typeof LoggerType
     ) {
         this.log = new Logger(__filename);
@@ -64,8 +68,6 @@ export class ListingItemTemplateService {
     @validate()
     public async create( @request(ListingItemTemplateCreateRequest) data: ListingItemTemplateCreateRequest): Promise<ListingItemTemplate> {
 
-        // this.log.debug('ListingItemTemplateService.create, data:', JSON.stringify(data, null, 2));
-
         const body = JSON.parse(JSON.stringify(data));
 
         // extract and remove related models from request
@@ -75,12 +77,8 @@ export class ListingItemTemplateService {
         delete body.paymentInformation;
         const messagingInformation = body.messagingInformation || [];
         delete body.messagingInformation;
+        const listingItemObjects = body.listingItemObjects || [];
         delete body.listingItemObjects;
-
-        // this.log.debug('itemInformation to save: ', JSON.stringify(itemInformation, null, 2));
-        // this.log.debug('paymentInformation to save: ', JSON.stringify(paymentInformation, null, 2));
-        // this.log.debug('messagingInformation to save: ', JSON.stringify(messagingInformation, null, 2));
-        // this.log.debug('listingItemObjects to save: ', JSON.stringify(listingItemObjects, null, 2));
 
         // If the request body was valid we will create the listingItemTemplate
         const listingItemTemplate: any = await this.listingItemTemplateRepo.create(body);
@@ -88,16 +86,20 @@ export class ListingItemTemplateService {
         if (!_.isEmpty(itemInformation)) {
             itemInformation.listing_item_template_id = listingItemTemplate.Id;
             const result = await this.itemInformationService.create(itemInformation as ItemInformationCreateRequest);
-            // this.log.debug('saved itemInformation ' + listingItemTemplate.Id + ': ', result.toJSON());
         }
         if (!_.isEmpty(paymentInformation)) {
             paymentInformation.listing_item_template_id = listingItemTemplate.Id;
             const result = await this.paymentInformationService.create(paymentInformation as PaymentInformationCreateRequest);
-            // this.log.info('saved paymentInformation: ', result.toJSON());
         }
         for (const msgInfo of messagingInformation) {
             msgInfo.listing_item_template_id = listingItemTemplate.Id;
             await this.messagingInformationService.create(msgInfo as MessagingInformationCreateRequest);
+        }
+
+        // create listingItemObjects
+        for (const object of listingItemObjects) {
+            object.listing_item_template_id = listingItemTemplate.Id;
+            await this.listingItemObjectService.create(object as ListingItemObjectCreateRequest);
         }
 
         // finally find and return the created listingItemTemplate
@@ -116,8 +118,6 @@ export class ListingItemTemplateService {
 
         // update listingItemTemplate record
         const updatedListingItemTemplate = await this.listingItemTemplateRepo.update(id, listingItemTemplate.toJSON());
-        this.log.debug('updatedListingItemTemplate.toJSON():', updatedListingItemTemplate.toJSON());
-
 
         // if the related one exists allready, then update. if it doesnt exist, create. and if the related one is missing, then remove.
         // Item-information
@@ -156,26 +156,72 @@ export class ListingItemTemplateService {
             await this.paymentInformationService.destroy(paymentInformation.id);
         }
 
-        // TODO: UPDATE!!
+        // // find related record and delete it and recreate related data
+        // let messagingInformation = updatedListingItemTemplate.related('MessagingInformation').toJSON() || [];
+        // for (const msgInfo of messagingInformation) {
+        //     msgInfo.listing_item_template_id = id;
+        //     await this.messagingInformationService.destroy(msgInfo.id);
+        // }
+        // // add new
+        // messagingInformation = body.messagingInformation || [];
+        // for (const msgInfo of messagingInformation) {
+        //     msgInfo.listing_item_template_id = id;
+        //     await this.messagingInformationService.create(msgInfo as MessagingInformationCreateRequest);
+        // }
 
         // find related record and delete it and recreate related data
+        const existintMessagingInformation = updatedListingItemTemplate.related('MessagingInformation').toJSON() || [];
 
-        // const paymentInformation = updatedListingItemTemplate.related('PaymentInformation').toJSON();
-        // await this.paymentInformationService.destroy(paymentInformation.id);
-        // body.paymentInformation.listing_item_template_id = id;
-        // await this.paymentInformationService.create(body.paymentInformation);
+        const newMessagingInformation = body.messagingInformation || [];
 
-        // find related record and delete it and recreate related data
-        let messagingInformation = updatedListingItemTemplate.related('MessagingInformation').toJSON() || [];
-        for (const msgInfo of messagingInformation) {
-            msgInfo.listing_item_template_id = id;
-            await this.messagingInformationService.destroy(msgInfo.id);
+        // delete MessagingInformation if not exist with new params
+        for (const msgInfo of existintMessagingInformation) {
+            if (!await this.checkExistingObject(newMessagingInformation, 'id', msgInfo.id)) {
+                await this.messagingInformationService.destroy(msgInfo.id);
+            }
         }
-        // add new
-        messagingInformation = body.messagingInformation || [];
-        for (const msgInfo of messagingInformation) {
+
+        // update or create messaging itemInformation
+        for (const msgInfo of newMessagingInformation) {
             msgInfo.listing_item_template_id = id;
-            await this.messagingInformationService.create(msgInfo as MessagingInformationCreateRequest);
+            const message = await this.checkExistingObject(existintMessagingInformation, 'id', msgInfo.id);
+            delete msgInfo.id;
+            if (message) {
+                message.protocol = msgInfo.protocol;
+                message.publicKey = msgInfo.publicKey;
+                await this.messagingInformationService.update(message.id, msgInfo as MessagingInformationUpdateRequest);
+            } else {
+                await this.messagingInformationService.create(msgInfo as MessagingInformationCreateRequest);
+            }
+        }
+
+        const newListingItemObjects = body.listingItemObjects || [];
+        // find related listingItemObjects
+        const existingListingItemObjects = updatedListingItemTemplate.related('ListingItemObjects').toJSON() || [];
+
+        // find highestOrderNumber
+        const highestOrderNumber = await this.findHighesOrderNumber(newListingItemObjects);
+
+        const objectsToBeUpdated = [] as any;
+        for (const object of existingListingItemObjects) {
+            // check if order number is greter than highestOrderNumber then delete
+            if (object.order > highestOrderNumber) {
+                await this.listingItemObjectService.destroy(object.id);
+            } else {
+                objectsToBeUpdated.push(object);
+            }
+        }
+
+        // create or update listingItemObjects
+        for (const object of newListingItemObjects) {
+            object.listing_item_template_id = id;
+            const itemObject = await this.checkExistingObject(objectsToBeUpdated, 'order', object.order);
+
+            if (itemObject) {
+                await this.listingItemObjectService.update(itemObject.id, object as ListingItemObjectUpdateRequest);
+            } else {
+                await this.listingItemObjectService.create(object as ListingItemObjectCreateRequest);
+            }
         }
 
         // finally find and return the updated listingItem
@@ -188,12 +234,6 @@ export class ListingItemTemplateService {
         const relatedCryptocurrencyAddress = listingItemTemplate
             .related('PaymentInformation')
             .toJSON();
-        // const relatedCryptocurrencyAddress = listingItemTemplate
-        //     .related('PaymentInformation')
-        //     .related('ItemPrice')
-        //     .related('CryptocurrencyAddress')
-        //     .toJSON();
-        // this.log.debug('relatedCryptocurrencyAddress: ', JSON.stringify(relatedCryptocurrencyAddress, null, 2));
 
         await this.listingItemTemplateRepo.destroy(id);
         // if we have cryptoaddress and it's not related to profile -> delete
@@ -203,4 +243,18 @@ export class ListingItemTemplateService {
 
     }
 
+    // check if object is exist in a array
+    private async checkExistingObject(objectArray: string[], fieldName: string, value: string | number): Promise<any> {
+        return await _.find(objectArray, (object) => {
+            return ( object[fieldName] === value );
+        });
+    }
+
+    // find highest order number from listingItemObjects
+    private async findHighesOrderNumber(listingItemObjects: string[]): Promise<any> {
+        const highestOrder = await _.maxBy(listingItemObjects, (itemObject) => {
+          return itemObject['order'];
+        });
+        return highestOrder ? highestOrder['order'] : 0;
+    }
 }
